+++ NEWS.md
@@ -26,6 +26,14 @@ This file collects a summary of important and/or user-visible changes.
       The bv output number format when printing the currently asserted input
       formula can now be configured via (new, required)
       parameter `uint8_t base`.
+    - `bitwuzla_parser_new(BitwuzlaOptions*, const char*, FILE*, const char*)`
+       changed to  
+      `bitwuzla_parser_new(BitwuzlaOptions*, const char*, FILE*, const char*,
+          uint8_t, const char*)`  
+      `BitwuzlaParser` is now configured with a bv output number format
+      (when printing model values) via (new, required) parameter `uint8_t base`,
+      and the name of the output file (`<stdout>` to use `stdout`) via (new,
+      required) parameter `const char* outfile_name`.
   * C++ API changes:
     - New stream modifier `set_bv_format(uint8_t)` allows to configure the
       output number format of bit-vector values of any output stream.